// types/phone.ts
export interface PhoneModel {
  id: number
  name: string
  image_url: string
  price: number
  colors: string
  display: string
  chip: string
  camera: string
  battery: string
  features: string
  release_year: number
  storage_options: string
  water_resistance: string
  dimensions: string
  weight: number
  series: string
}